assignment-2-784478/code/fetchdata/requirements.txt
apscheduler
pyyaml
watchdog